chore(monorepo): update README and CONTIRBUTING
After several questions from the community, it would be a good idea
to have the getting started in the README itself, instead of hidden
in the CONTRIBUTION guide.

After comments on issue resolution in the monorepo, it is reasonable
to have a basic guide on how to fix problems. Relevant steps have
been added to the CONTRIBUTION guide.